Move autoalloc screen to a separate module
