+++ crates/hyperqueue/src/dashboard/ui/screens/autoalloc/mod.rs
