+++ crates/hyperqueue/src/dashboard/ui/screens/mod.rs
@@ -1,4 +1,4 @@
-pub mod autoalloc_screen;
+pub mod autoalloc;
 pub mod cluster;
 pub mod jobs;
 pub mod root_screen;
